requirements.txt
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asteval==0.9.29
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
cssselect==1.2.0
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==0.8.3
fake-useragent==1.1.3
fastjsonschema==2.16.1
feedparser==6.0.10
fonttools==4.34.4
frozendict==2.3.4
future==0.18.3
h11==0.14.0
html5lib==1.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
lmfit==1.1.0
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
multitasking==0.0.11
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
newsapi-python==0.2.7
notebook==6.4.12
numpy==1.23.1
opencv-python==4.6.0.66
openpyxl==3.1.0
outcome==1.2.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
progressbar==2.5
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyee==8.2.2
Pygments==2.12.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==2.0.0
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==23.2.0
requests==2.28.2
requests-html==0.10.0
reticker==3.1.6
runipy==0.1.5
scipy==1.8.1
selenium==4.9.1
Send2Trash==1.8.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.3.0
statsmodels==0.13.5
stocksymbol==0.0.5
terminado==0.15.0
tinycss2==1.1.1
tornado==6.2
tqdm==4.65.0
traitlets==5.3.0
trio==0.22.0
trio-websocket==0.10.2
uncertainties==3.1.7
Unidecode==1.3.6
urllib3==1.26.14
w3lib==2.1.1
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.4
wget==3.2
wsproto==1.2.0
XlsxWriter==3.0.9
yahoo-fin==0.8.9.1
yfinance==0.2.4
zipp==3.15.0
